% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cars}
\alias{cars}
\title{cars data set}
\format{A zooreg object with 93 rows and 3 variables:}
\source{
downloaded with sophisthse("HHI_Q_I") command
}
\usage{
cars
}
\description{
xxx
}
\keyword{datasets}

