% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cities}
\alias{cities}
\title{Russian cities with population of over 1 million}
\format{A SpatialPointsDataFrame object with 15 points with following attibutes:
\describe{
 \item{id}{custom id assigned}
 \item{name}{city name}
 \item{pop_2015}{population in 2015, according to Rosstat}
 }}
\source{
Obtained with QGIS using OpenStreetMap data. Data tables were customized. Population data retrieved from Rosstat. MapData©OpenStreetMap
}
\usage{
cities
}
\description{
The dataset contains names, population and coordinates of Russian cities with population greater than 1 million.
}
\keyword{datasets}

