% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Khabarovsk}
\alias{Khabarovsk}
\title{Khabarovsk region municipal districts and Khabarovsk city districts}
\format{A SpatialPolygonsDataFrame object with 24 polygons with following attibutes:
\describe{
 \item{id}{custom id assigned}
 \item{name}{region name}
 \item{admin_lvl}{administrative level of the region}
 \item{pop_2015}{population in 2015, according to Rosstat}
 \item{pop_2016}{population in 2016, according to Rosstat}
 }}
\source{
Obtained from GISLAB. QGIS was used to simplify geometry and to customize data table. Population data retrieved from Rosstat. MapData©OpenStreetMap
}
\usage{
Khabarovsk
}
\description{
The dataset contains names, population (as of 2015 and 2016) and administrative level of included regions.
}
\keyword{datasets}

