% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rusmaps.R
\name{longitude_correction}
\alias{longitude_correction}
\title{Set positive values to negative longitudes}
\usage{
longitude_correction(x)
}
\arguments{
\item{x}{numeric vector of longitudes}
}
\value{
non-negative versions of longitued
Non-negative versions of longitudes
}
\description{
longitude_correction() receives a vector of longitudes
and turns them positive in case they are not.
}
\examples{
longitude_correction(c(-3.5, 10.8))
}

