% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Novosibirsk}
\alias{Novosibirsk}
\title{Administrative regions of Novosibirsk city}
\format{A SpatialPolygonsDataFrame object with 10 polygons with following attibutes:
\describe{
 \item{id}{custom id assigned}
 \item{name}{region name}
 \item{admin_lvl}{administrative level of the region}
 \item{pop_2016}{population in 2016, according to Rosstat}
 }}
\source{
Obtained with QGIS using OpenStreetMap data. Data tables were customized. Population data retrieved from Rosstat. MapData(c)OpenStreetMap
}
\usage{
Novosibirsk
}
\description{
The dataset contains names, population (as of 2016) and administrative level of included regions.
}
\keyword{datasets}

