% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Omsk}
\alias{Omsk}
\title{Administrative regions of Omsk city}
\format{A SpatialPointsDataFrame object with 5 polygons with following attibutes:
\describe{
 \item{id}{custom id assigned}
 \item{name}{region name}
 \item{admin_lvl}{administrative level of the region}
 \item{pop_2010}{population in 2010, according to Rosstat}
 }}
\source{
Obtained with QGIS using OpenStreetMap data. Data tables were customized. Population data retrieved from Rosstat.
}
\usage{
Omsk
}
\description{
The dataset contains names, population (as of 2010) and administrative level of included regions
}
\keyword{datasets}

