% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rusmaps.R
\name{plus_rail}
\alias{plus_rail}
\title{It's square a number}
\usage{
plus_rail(mo_f)
}
\arguments{
\item{mo_f}{a number which will be squared}
}
\value{
The square of x
}
\description{
It's square a number
}
\details{
Build map with railways
}
\examples{
square(-1)
}

