% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rusmaps.R
\name{restore180}
\alias{restore180}
\title{Restore a polygon cut by 180 meridian}
\usage{
restore180(y)
}
\arguments{
\item{y}{a SpatialPolygonsDataFrame object}
}
\value{
Whole polygons, if they were cut by 180 meredian.
}
\description{
The function reads a file of SpatialPolygonsDataFrame and unifies
polygons if they are separated by 180 meridian
}
\examples{
restore180(cities)
}

