% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rusmaps.R
\name{longitude_correction}
\alias{longitude_correction}
\title{longitude_correction (brief)}
\usage{
longitude_correction(x)
}
\arguments{
\item{x}{numeric vector of longitudes}
}
\value{
non-negative versions of longitued
}
\description{
longitude_correction (more detailed)
}
\details{
longitude_correction (even more detailed)
}
\examples{
longitude_correction(c(-3.5, 10.8))
}

