% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rusmaps.dataframe}
\alias{rusmaps.dataframe}
\title{Maps in rusmaps package}
\format{A DataFrame object with 30 rows and following attibutes:
\describe{
 \item{map_name}{name of dataset containing a map}
 \item{description}{what the dataset contains}
 \item{format}{format of dataset}
 \item{date}{relevance date}
 }}
\source{
Compiled from existing data.
}
\usage{
rusmaps.dataframe
}
\description{
The dataset contains map names (how they can be called), a brief descrtiption, format and relevance date.
}
\keyword{datasets}

